collector-python/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pika==1.3.2
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
pandas==2.2.2
numpy==1.26.4
